Move button into directives folder
