+++ angular/projects/spark-angular/src/lib/directives/sprk-button/sprk-button.module.ts
@@ -6,6 +6,6 @@ import { SprkButtonDirective } from './sprk-button.directive';
 @NgModule({
   imports: [CommonModule, RouterModule],
   declarations: [SprkButtonDirective],
-  exports: [SprkButtonDirective]
+  exports: [SprkButtonDirective],
 })
 export class SprkButtonModule {}
